+++ Core/PixelEvent.swift
@@ -361,6 +361,8 @@ extension Pixel {
         
         case debugCantSaveBookmarkFix
         
+        case debugDailyPixelStorageError
+        
         // Errors from Bookmarks Module
         case bookmarkFolderExpected
         case bookmarksListIndexNotMatchingBookmark
@@ -717,6 +719,8 @@ extension Pixel.Event {
         case .debugMissingTopFolderFixHasFavorites: return "m_d_missing_top_folder_has_favorites"
             
         case .debugCantSaveBookmarkFix: return "m_d_cant_save_bookmark_fix"
+            
+        case .debugDailyPixelStorageError: return "m_d_daily_pixel_storage_error"
             
         
         // MARK: Ad Attribution
