+++ Core/TimedPixel.swift
@@ -1,0 +1,39 @@
+//
+//  TimedPixel.swift
+//  DuckDuckGo
+//
+//  Copyright © 2023 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import Foundation
+
+public class TimedPixel {
+    
+    let pixel: Pixel.Event
+    let date: Date
+    
+    public init(_ pixel: Pixel.Event, date: Date = Date()) {
+        self.pixel = pixel
+        self.date = date
+    }
+    
+    public func fire(_ fireDate: Date = Date(), withAdditionalParameters params: [String: String] = [:]) {
+        let duration = String(fireDate.timeIntervalSince(date))
+        var newParams = params
+        newParams[PixelParameters.duration] = duration
+        Pixel.fire(pixel: pixel, withAdditionalParameters: newParams)
+    }
+    
+}
